add 3 files(js/css/icon) into static, add regex to verify member account, adjust homepage layout(search box), add .env
