+++ templates/thankyou.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>付款完成</title>
-    <link href="../static/thankyou.css" rel="stylesheet" type="text/css" media="all" />
-    <link href="../static/memberForm.css" rel="stylesheet" type="text/css" media="all" />
+    <link href="../static/css/thankyou.css" rel="stylesheet" type="text/css" media="all" />
+    <link href="../static/css/memberForm.css" rel="stylesheet" type="text/css" media="all" />
 </head>
 <body onload="userStatus();">
     <section id="fixed">
@@ -24,9 +24,8 @@
     <div class="end" id="end">
         <div id="copyright">COPYRIGHT © 2021 台北一日遊</div>
     </div>
-    <script src="../static/thankyou.js"></script>
-    <script src="../static/booking.js"></script>
-    <script src="../static/memberForm.js"></script>
-    <script src="../static/memberCheck.js"></script>
+    <script src="../static/js/thankyou.js"></script>
+    <script src="../static/js/memberForm.js"></script>
+    <script src="../static/js/memberCheck.js"></script>
 </body>
 </html>
